compute/etl-engine/pyproject.toml
[project]
name = "etl-engine"
version = "0.1.0"
description = "ETL execution engine for Mellivora Mind Studio"
requires-python = ">=3.11"
dependencies = [
    # Data processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyarrow>=14.0.0",
    
    # Database
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "clickhouse-driver>=0.2.6",
    "redis>=5.0.0",
    
    # Data sources
    "tushare>=1.2.89",
    "requests>=2.31.0",
    
    # Scheduling
    "apscheduler>=3.10.0",
    "croniter>=2.0.0",
    
    # gRPC
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    
    # Async
    "asyncpg>=0.29.0",
    
    # Web server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    
    # Utils
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.11"
strict = true
